cleanup

+++ tests/day05.rs
@@ -5,7 +5,7 @@ mod tools;
 fn test_part1(){
     let input = tools::import_test_input(5);
 
-    let expected:u32 = 12;
+    let expected:u64 = 12;
     let result = find_overlaps(input, 10,true);
     assert_eq!(expected, result)
 }
