+++ tests/day06.rs
@@ -5,10 +5,10 @@ use advent_of_code::day06::sub_modular;
 fn test_day06(){
     let input: String = String::from("3,4,3,1,2");
     let result = populate_laternfish(&input, 18);
-    let expected: u32 = 26;
+    let expected: u64 = 26;
     assert_eq!(expected, result);
     let result = populate_laternfish(&input, 80);
-    let expected: u32 = 5934;
+    let expected: u64 = 5934;
     assert_eq!(expected, result)
 }
