+++ tests/day07.rs
@@ -4,9 +4,9 @@ use advent_of_code::day07::find_optimal_position;
 fn test_day07(){
     let input: String = String::from("16,1,2,0,4,2,7,1,2,14");
     let result = find_optimal_position(&input, false);
-    let expected: u32 = 37;
+    let expected: u64 = 37;
     assert_eq!(expected, result);
     let result = find_optimal_position(&input, true);
-    let expected: u32 = 168;
+    let expected: u64 = 168;
     assert_eq!(expected, result);
 }
