+++ tests/day09.rs
@@ -6,7 +6,7 @@ mod tools;
 fn test_day09_p1() {
     let input = tools::import_test_input(9);
     let result = find_hotspots(input, 10);
-    let expected: u32 = 15;
+    let expected: u64 = 15;
     assert_eq!(expected, result);
 }
 
@@ -14,6 +14,6 @@ fn test_day09_p1() {
 fn test_day09_p2() {
     let input = tools::import_test_input(9);
     let result = find_basins(input, 10);
-    let expected: u32 = 1134;
+    let expected: u64 = 1134;
     assert_eq!(expected, result);
 }
